pyproject.toml
[project]
name = "fastapi-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "httpx>=0.24.0",
    "jinja2>=3.1.0",
    "python-jose[cryptography]>=3.5.0",
    "passlib[bcrypt]>=1.7.0",
    "uvicorn>=0.30.0",
    "pydantic[email]>=2.0.0",
    "python-dotenv>=1.0.0",
    "websockets>=15.0.1",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.21.0",
]

[tool.setuptools]
packages = ["app"]
